Migrate all serverside and common code to typescript, let build .js files be built successfully using commonjs modules for now

+++ master.ts
@@ -1,43 +1,57 @@
-const cluster = require("cluster");
-const http = require("http");
-const numCPUs = require("os").cpus().length;
-const url = require("url");
-const express = require("express");
-const PORT = process.env.PORT || 3000;
-const path = require("path");
-const fs = require("fs");
-const { v4: uuidv4 } = require("uuid");
+import http from "http";
+import cluster from "cluster";
+import os from "os";
 
-const cors = require("cors");
-const nbLoop = require("./common/nonBlockingLoop");
+import express, { Response } from "express";
+import path from "path";
+import fs from "fs";
+import { v4 as uuidv4 } from "uuid";
+import cors from "cors";
+import { Worker } from "cluster";
+const PORT = process.env.PORT || 3000;
+const numCPUs = os.cpus().length;
 
 const app = express();
 app.use(cors());
 app.use(express.static("dist"));
 app.use(express.static("public"));
 
-var clusterWorkersPort = null;
-const WorkerIdToPendingHTTPRequest = {};
+var clusterWorkersPort: number | null = null;
+const WorkerIdToPendingHTTPRequest: {
+  [key: number]: Response;
+} = {};
 
 class SessionManager {
+  sessions: {
+    [key: string]: {
+      sessionId: string;
+      workerId: number;
+      open: boolean;
+      playerCount: number;
+    };
+  };
   constructor() {
     this.sessions = {};
   }
-  getSessionsByWorkerId(workerId) {
+  getSessionsByWorkerId(workerId: number) {
     if (!workerId) return null;
     let sessionsByWorkerId = Object.values(this.sessions).filter(
       (sessionData) => sessionData.workerId === workerId
     );
     return sessionsByWorkerId;
   }
-  getSessions(sid = null) {
-    if (sid === null) return Object.values(this.sessions);
-    return this.sessions[sid] || null;
+  getSessions(sid?: string) {
+    if (!sid) return Object.values(this.sessions);
+    return [this.sessions[sid]] || null;
   }
-  killSession(sid) {
+  killSession(sid: string) {
     if (sid && this.sessions.hasOwnProperty(sid)) delete this.sessions[sid];
   }
-  createNewSession(sid, workerId, isAcceptingNewConnections = true) {
+  createNewSession(
+    sid: string,
+    workerId: number,
+    isAcceptingNewConnections = true
+  ) {
     this.sessions[sid] = {
       sessionId: sid,
       workerId,
@@ -45,7 +59,10 @@ class SessionManager {
       playerCount: 0,
     };
   }
-  updateSession(sid, updateParams) {
+  updateSession(
+    sid: string,
+    updateParams: { open?: boolean; playerCount?: number }
+  ) {
     const acceptedParams = ["playerCount", "open"];
     if (typeof updateParams !== "object") {
       throw new Error("updateParams must be an object.");
@@ -67,31 +84,37 @@ class SessionManager {
 }
 
 class WorkerManager {
-  constructor(maxSessionsPerWorker) {
-    console.log('maxSessionsPerWorker', maxSessionsPerWorker);
+  workers: {
+    [key: string]: {
+      worker: Worker;
+      sessionManager: SessionManager;
+    };
+  };
+  MAX_SESSIONS_PER_WORKER: number;
+  constructor(maxSessionsPerWorker: number) {
+    console.log("maxSessionsPerWorker", maxSessionsPerWorker);
     this.workers = {}; //workerId as a key
     this.MAX_SESSIONS_PER_WORKER = Number(maxSessionsPerWorker);
   }
 
   //Returns worker which is capable of hosting a new session.
   getAvailableWorker() {
-    let availableWorker = null;
-    availableWorker =
-      Object.values(this.workers).find(
-        (workerData) =>
-          workerData.sessionManager.getSessions().length <
-          this.MAX_SESSIONS_PER_WORKER
-      )?.worker || null;
-    return availableWorker;
+    let availableWorker;
+    availableWorker = Object.values(this.workers).find(
+      (workerData) =>
+        workerData.sessionManager.getSessions().length <
+        this.MAX_SESSIONS_PER_WORKER
+    )?.worker;
+    return availableWorker || null;
   }
 
   //Returns array of all workers if no workerId provided,
   //otherwise returns a single worker-data object.
   getWorkers(workerId = null) {
-    if(!workerId) {
+    if (!workerId) {
       return Object.values(this.workers);
     }
-    return this.workers[workerId] || null;
+    return [this.workers[workerId]] || null;
   }
 
   //Creates a new worker process with a session.
@@ -105,11 +128,9 @@ class WorkerManager {
     this.createSession(workerInstance.id);
     return this.workers[workerInstance.id];
   }
-  createSession(wid) {
+  createSession(wid: number) {
     const sessionID = uuidv4();
-    this.workers[wid]
-    .sessionManager
-    .createNewSession(sessionID, wid);
+    this.workers[wid].sessionManager.createNewSession(sessionID, wid);
 
     this.workers[wid].worker.send({
       type: "SESSION_CREATE_REQUESTED",
@@ -118,37 +139,42 @@ class WorkerManager {
     });
     return sessionID;
   }
-  getSessionManager(wid) {
+  getSessionManager(wid: number) {
     return this.workers[wid].sessionManager || null;
   }
   getSessions(wid = null) {
-    if(!wid) {
-      let allSessions = Object.values(this.workers).map(({sessionManager}) => sessionManager.getSessions()).flat(1);
+    if (!wid) {
+      let allSessions = Object.values(this.workers)
+        .map(({ sessionManager }) => sessionManager.getSessions())
+        .flat(1);
       return allSessions;
     }
     let sessions = this.workers[wid].sessionManager.getSessions();
     return sessions;
   }
-  getWorkerBySessionId(sessionId) {
-    let worker = Object.values(this.workers).find(worker => {
-      worker.sessionManager.getSessions(sessionId) !== null
+  getWorkerBySessionId(sessionId: string) {
+    let worker = Object.values(this.workers).find((worker) => {
+      worker.sessionManager.getSessions(sessionId) !== null;
     });
-    if(!worker) return null;
+    if (!worker) return null;
     return worker;
   }
-  killWorker(workerId) {
-    if(!this.workers.hasOwnProperty(workerId)) {
-      console.log(`[WorkerManager/killWorker]: Failed to find worker #${workerId}, either already killed or invalid ID.`);
+  killWorker(workerId: number) {
+    if (!this.workers.hasOwnProperty(workerId)) {
+      console.log(
+        `[WorkerManager/killWorker]: Failed to find worker #${workerId}, either already killed or invalid ID.`
+      );
       return;
     }
     const isDead = this.workers[workerId].worker.isDead();
     console.log(`[WorkerManager/killWorker]: ${workerId} (isDead: ${isDead}) `);
-    if(!isDead)
-      this.workers[workerId].worker.kill();
+    if (!isDead) this.workers[workerId].worker.kill();
     delete this.workers[workerId];
   }
 }
-const workerManager = new WorkerManager(Number(process.env.MAX_SESSION_PER_WORKER));
+const workerManager = new WorkerManager(
+  Number(process.env.MAX_SESSION_PER_WORKER)
+);
 
 const server = http.createServer(app);
 
@@ -188,38 +214,50 @@ app.post("/session", async (req, res) => {
         message: `Unable to create session, max capacity reached on this instance.`,
         code: "MAX_SESSIONS_REACHED",
       });
-    
+
     //creates a worker with 1 session.
-    workerData = workerManager.createWorker();
+    const workerData = workerManager.createWorker();
     availableWorker = workerData.worker;
-  }
-  else {
+  } else {
     workerManager.createSession(availableWorker.id);
   }
   WorkerIdToPendingHTTPRequest[availableWorker.id] = res;
+  console.log(`stored res for ${availableWorker.id}`)
 });
 
 app.get("/sessions", (req, res) => {
   if (!clusterWorkersPort) return res.status(500).send({});
-  let { id: sessionId, limit, offset } = req.query;
+  let {
+    id: sessionId,
+    limit,
+    offset,
+  }: { id: string; limit: number; offset: number } = req.query as unknown as {
+    id: string;
+    limit: number;
+    offset: number;
+  };
   limit = limit || 1;
   offset = offset || 0;
-  
-  if(sessionId) {
+
+  if (sessionId) {
     let fetchedSession = workerManager
       .getWorkerBySessionId(sessionId)
-      ?.sessionManager?.getSessions(sessionId);
-    
-      if (!fetchedSession)
-        return res.status(404).send();
-      return res.status(200).json({
-        sessions: [fetchedSession.sessionId],
-        port: clusterWorkersPort
-      });
+      ?.sessionManager?.getSessions(sessionId)
+      ?.at(0);
+
+    if (!fetchedSession) return res.status(404).send();
+
+    return res.status(200).json({
+      sessions: [fetchedSession.sessionId],
+      port: clusterWorkersPort,
+    });
   }
 
   // fetch all available sessions
-  let availableSessions = workerManager.getSessions().filter(s => s.open).map(s => s.sessionId);
+  let availableSessions = workerManager
+    .getSessions()
+    .filter((s) => s.open)
+    .map((s) => s.sessionId);
   return res.status(200).json({
     sessions: availableSessions,
     port: clusterWorkersPort,
@@ -232,13 +270,17 @@ app.get("*", (req, res) => {
 
 //When any of the workers die
 cluster.on("exit", (worker, code, signal) => {
-  console.log(`[cluster exit]: Worker #${worker.id} exited | [code:${code}] [signal:${signal}]`);
+  console.log(
+    `[cluster exit]: Worker #${worker.id} exited | [code:${code}] [signal:${signal}]`
+  );
   workerManager.killWorker(worker.id);
 });
 
 //Emitted after the worker IPC channel has disconnected
 cluster.on("disconnect", (worker) => {
-  console.log(`[cluster-disconnect]: Worker#${worker.id} disconnected: ${worker.isDead()}`);
+  console.log(
+    `[cluster-disconnect]: Worker#${worker.id} disconnected: ${worker.isDead()}`
+  );
   workerManager.killWorker(worker.id);
   if (WorkerIdToPendingHTTPRequest.hasOwnProperty(worker.id)) {
     WorkerIdToPendingHTTPRequest[worker.id].status(503).json({
@@ -249,18 +291,17 @@ cluster.on("disconnect", (worker) => {
 });
 
 cluster.on("message", (worker, message) => {
-  console.log('cluster message received : ', message);
+  console.log("cluster message received : ", message);
   if (message.type === "SESSION_CREATED_ACK") {
     const { sessionId, workerId } = message;
     const responseData = {
       sessionId: sessionId,
       port: clusterWorkersPort,
     };
-
+    
     WorkerIdToPendingHTTPRequest[worker.id].status(200).json(responseData);
     delete WorkerIdToPendingHTTPRequest[worker.id];
-  } 
-  else if (message.type === "SESSION_UPDATED") {
+  } else if (message.type === "SESSION_UPDATED") {
     const { sessionId, gameStarted, players } = message;
     workerManager
       .getWorkerBySessionId(sessionId)
@@ -268,10 +309,11 @@ cluster.on("message", (worker, message) => {
         open: !gameStarted,
         playerCount: players,
       });
-  } 
-  else if (message.type === "SESSION_DESTROYED") {
+  } else if (message.type === "SESSION_DESTROYED") {
     const { sessionId, workerId } = message;
-    console.log(`master: Received SESSION_DESTROYED: Killing session(${sessionId})`);
+    console.log(
+      `master: Received SESSION_DESTROYED: Killing session(${sessionId})`
+    );
     workerManager.getSessionManager(workerId).killSession(sessionId);
   }
 });
@@ -284,5 +326,7 @@ cluster.on("listening", (worker, address) => {
 });
 
 server.listen(PORT, () => {
-  console.log(`Master HTTP Server listening on port ${server.address().port}`);
+  console.log(
+    `Master HTTP Server listening on port ${(server.address() as any).port}`
+  );
 });
